requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Deprecated==1.2.14
distro==1.9.0
Flask==3.0.2
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.14.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-asyncio==0.44b0
opentelemetry-instrumentation-aws-lambda==0.44b0
opentelemetry-instrumentation-dbapi==0.44b0
opentelemetry-instrumentation-flask==0.44b0
opentelemetry-instrumentation-grpc==0.44b0
opentelemetry-instrumentation-httpx==0.44b0
opentelemetry-instrumentation-jinja2==0.44b0
opentelemetry-instrumentation-logging==0.44b0
opentelemetry-instrumentation-openai==0.14.1
opentelemetry-instrumentation-sqlite3==0.44b0
opentelemetry-instrumentation-system-metrics==0.44b0
opentelemetry-instrumentation-tortoiseorm==0.44b0
opentelemetry-instrumentation-urllib==0.44b0
opentelemetry-instrumentation-wsgi==0.44b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-propagator-b3==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-semantic-conventions-ai==0.0.23
opentelemetry-test-utils==0.44b0
opentelemetry-util-http==0.44b0
packaging==24.0
protobuf==4.25.3
psutil==5.9.8
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
sniffio==1.3.1
splunk-opentelemetry==1.18.0
tqdm==4.66.2
typing_extensions==4.10.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.0
